fix: move enum to constants
