+++ packages/sdk/src/chat/constant.ts
@@ -1,0 +1,7 @@
+export const InstillChatMessageEnum = {
+  Started: "CHAT_STARTED",
+  StatusUpdated: "CHAT_STATUS_UPDATED",
+  OutputUpdated: "CHAT_OUTPUT_UPDATED",
+  Ended: "CHAT_ENDED",
+  UserMessage: "CHAT_USER_MESSAGE",
+} as const;
