+++ packages/sdk/src/chat/index.ts
@@ -1,2 +1,3 @@
 export * from "./ChatClient";
 export * from "./types";
+export * from "./constant";
